fix: stabilize runtime worker lifecycle paths
